fixed TypeError for slice indexing
This fixes a "TypeError: slice indices must be integers or None or have
an __index__ method" because a non-integer index was provided.

+++ src/dehaze.py
@@ -55,7 +55,7 @@ def get_atmosphere(I, darkch, p):
     M, N = darkch.shape
     flatI = I.reshape(M * N, 3)
     flatdark = darkch.ravel()
-    searchidx = (-flatdark).argsort()[:M * N * p]  # find top M * N * p indexes
+    searchidx = (-flatdark).argsort()[:int(M * N * p)]  # find top M * N * p indexes
     print 'atmosphere light region:', [(i / N, i % N) for i in searchidx]
 
     # return the highest intensity for each channel
